pyproject.toml
[project]
name = "gcp-compute-inv"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "dotenv>=0.9.9",
    "google-api-python-client>=2.170.0",
    "google-auth>=2.39.0",
    "google-cloud>=0.34.0",
    "google-cloud-compute>=1.30.0",
    "google-cloud-container>=2.56.1",
    "google-cloud-resource-manager>=1.14.2",
    "pytz>=2025.2",
    "redis>=6.0.0",
    "structlog>=25.3.0",
]
